pyproject.toml
[tool.poetry]
name = "coolify-docker-converter"
version = "0.1.0"
description = "Takes ENV files and converts them to a Coolify template only needing descriptions and labels, can optionally take YAML's for more structured Coolify templates"
authors = ["Zachary Handley <zachhandley@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "^6.0.1"
pydantic = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
